+++ monitoring/src/monitoring_service/constants.py
@@ -1,1 +1,2 @@
 WAIT_BLOCKS_BEFORE_MONITOR: int = 100
+DEFAULT_REQUIRED_CONFIRMATIONS: int = 10
